+++ lib/src/migrations/interfaces/csm_set_interface.dart
@@ -1,0 +1,16 @@
+import 'package:csm_foundation_services/csm_foundation_services.dart';
+import 'package:csm_foundation_services/src/migrations/structures/csm_set_validation_result.dart';
+
+/// Defines a [Migration] internal product concept interface to define
+/// base requirements for [Set] implementations and definitions.
+///
+/// [Migration]: concept related to everything from the databases sources.
+abstract interface class CSMSetInterface implements CSMEncodeInterface {
+  /// Set record pointer.
+  late int id;
+
+  /// Client level evaluation for this set record to check if can be written correctly by the service.
+  ///
+  /// If the result list came empty, means no validation results were thrown (meaning the evaluation is correct).
+  List<CSMSetValidationResult> evaluate();
+}