+++ test/tools/csm_deserealizer/csm_deserealizer_web_test.dart
@@ -1,4 +1,6 @@
 @TestOn('browser')
+library;
+
 
 import 'package:csm_foundation_services/csm_foundation_services.dart';
 import 'package:test/test.dart';